refactor(domain): split large test files by concern
